Be more picky about hash functions by default
RIPEMD160 is an older function, about the same age as MD5, so we don't want to start using it in new code.

SHA-224 isn't one that we use much. It's more or less a truncated form of SHA-256, and SHA-256 itself works in places where SHA-224 would, so we opt for the longer output of SHA-256.