Remove references to html built via string operations
While it's certainly possible to carefully construct HTML fragments with careful use of `html_safe`, we suggest that string operations should not be shown in examples where HTML is built. The safer approach to constructing HTML would be to use templates or tag helpers that automatically handle the context-aware and `html_safe`-aware APIs. 

So I took the approach of removing all HTML references to focus on what the rules are actually demonstrating. I could see an argument for converting the examples to use tag helpers but that would probably change the form and intent of these examples.