Remove version constraints for rubocop libraries
These constraints make it more difficult to upgrade these libraries in
projects that use rubocop-github.

I think the idea here was that we'd be able to bump the version
alongside the default config, but without locking to an exact version we
can still end up with a config that is incompatible with certain rubocop
version (see the unrecognized cop error in
https://github.com/github/rubocop-github/pull/86, for example).

If anything, I'd prefer a `>=` constraint over a `<=`. That would allow
us to bump the minimum constraint when we add new cops, but still
freely upgrade in our projects as long as there are no config
incompatibilities (which are less common anyway ever since RuboCop 1.0).