Bump all dev dependencies
This commit bumps all development dependencies, and removes pessimistic
version constraints where possible now that we've got a Gemfile.lock checked
in.

We had to keep the constraint for Rails for now, since Rails 7 is not
compatible with Ruby 2.6. We can bump Rails if we drop Ruby 2.6.

None of this affects the released gem, so as long as tests still pass we
are in good shape.