Revert "Run `bundle exec rubocop -a` (or `-A`) on all the code in this repo"
This reverts commit 53a0a37784c3d95c05edbc3f57023072b00ef607.

This is because we've now disabled all the rules that would have been
`DisabledByDefault`. This is so that downstream consumers don't get any
nasty surprises. As a result, there should be no RuboCop linting
violations in the code in this repo, so we can back out the changes to
slim the diff: they're now redundant.

Of course, it would be a good idea to go and make these changes at a later
date, but we can do that when we enable the correct rules having matched
each one up with its rationale in our styleguide.