Revert "Fix `Naming/MemoizedInstanceVariableName` offenses"
This reverts commit e36d683f6b85e7394775a02d9f6321f2cf34b4b5.

 This is because we've now disabled all the rules that would have been
`DisabledByDefault`. This is so that downstream consumers don't get any
nasty surprises. As a result, there should be no RuboCop linting
violations in the code in this repo, so we can back out the changes to
slim the diff: they're now redundant.

Of course, it would be a good idea to go and make these changes at a later
date, but we can do that when we enable the correct rules having matched
each one up with its rationale in our styleguide.