Specify the plugin class name explicitly
This allows rubocop to be run in bundler standalone mode. Without
specifying the plugin class name, rubocop errors out with:

> Error: Failed to load plugin `rubocop-performance` because the corresponding plugin class could not be determined for instantiation.

This is because it's trying to load the gem to find the name, but is
unable to.

Setting the class name doesn't affect running rubocop through bundler,
it will still work properly